refactor getting previewApiKey in pages
