--- lib/utils/routeParams.ts
@@ -1,8 +1,0 @@
-
-export const getEnvIdFromRouteParams = (envId: string | undefined) => {
-  if(!envId) {
-    throw new Error("The envId route parameter is missing");
-  }
-
-  return envId;
-}
