+++ pages/[envId]/index.tsx
@@ -9,7 +9,7 @@ import { getHomepage, getSiteMenu } from '../../lib/kontentClient';
 import { ValidCollectionCodename } from '../../lib/types/perCollection';
 import { useSmartLink } from '../../lib/useSmartLink';
 import { defaultEnvId, siteCodename } from '../../lib/utils/env';
-import { getEnvIdFromRouteParams } from '../../lib/utils/routeParams';
+import { getEnvIdFromRouteParams, getPreviewApiKeyFromPreviewData } from '../../lib/utils/pageUtils';
 import { Nav_NavigationItem, WSL_WebSpotlightRoot } from '../../models';
 
 
@@ -65,9 +65,7 @@ const Home: NextPage<Props> = props => {
 export const getStaticProps: GetStaticProps<Props, { envId: string }> = async context => {
   const envId = getEnvIdFromRouteParams(context.params?.envId);
 
-  const previewApiKey = context.previewData && typeof context.previewData === 'object' && 'currentPreviewApiKey' in context.previewData
-    ? context.previewData.currentPreviewApiKey as string
-    : undefined;
+  const previewApiKey = getPreviewApiKeyFromPreviewData(context.previewData);
 
   const homepage = await getHomepage({ envId, previewApiKey }, !!context.preview);
   const siteMenu = await getSiteMenu({ envId, previewApiKey }, !!context.preview);
