+++ pages/[envId]/products/[slug].tsx
@@ -8,7 +8,7 @@ import { mainColorButtonClass, mainColorHoverClass, mainColorTextClass } from ".
 import { getDefaultMetadata, getProductDetail, getProductItemsWithSlugs, getSiteMenu } from "../../../lib/kontentClient";
 import { ValidCollectionCodename } from "../../../lib/types/perCollection";
 import { defaultEnvId, siteCodename } from "../../../lib/utils/env";
-import { getEnvIdFromRouteParams } from "../../../lib/utils/routeParams";
+import { getEnvIdFromRouteParams, getPreviewApiKeyFromPreviewData } from "../../../lib/utils/pageUtils";
 import { createElementSmartLink } from "../../../lib/utils/smartLinkUtils";
 import { contentTypes, Metadata, Nav_NavigationItem, Product } from "../../../models";
 
@@ -47,9 +47,7 @@ export const getStaticProps: GetStaticProps<Props, IParams> = async (context) =>
 
   const envId = getEnvIdFromRouteParams(context.params?.envId);
 
-  const previewApiKey = context.previewData && typeof context.previewData === 'object' && 'currentPreviewApiKey' in context.previewData
-    ? context.previewData.currentPreviewApiKey as string
-    : undefined;
+  const previewApiKey = getPreviewApiKeyFromPreviewData(context.previewData);
 
   const product = await getProductDetail({ envId, previewApiKey }, slug, !!context.preview);
   const siteMenu = await getSiteMenu({ envId, previewApiKey }, !!context.preview);
