+++ pages/[envId]/solutions/[slug].tsx
@@ -9,7 +9,7 @@ import { mainColorBgClass } from "../../../lib/constants/colors";
 import { getDefaultMetadata, getSiteMenu, getSolutionDetail, getSolutionsWithSlugs } from "../../../lib/kontentClient";
 import { ValidCollectionCodename } from "../../../lib/types/perCollection";
 import { defaultEnvId, siteCodename } from "../../../lib/utils/env";
-import { getEnvIdFromRouteParams } from "../../../lib/utils/routeParams";
+import { getEnvIdFromRouteParams, getPreviewApiKeyFromPreviewData } from "../../../lib/utils/pageUtils";
 import { createElementSmartLink } from "../../../lib/utils/smartLinkUtils";
 import { contentTypes, Metadata, Nav_NavigationItem, Solution } from "../../../models";
 
@@ -50,9 +50,7 @@ export const getStaticProps: GetStaticProps<Props, IParams> = async (
 
   const envId = getEnvIdFromRouteParams(context.params?.envId);
 
-  const previewApiKey = context.previewData && typeof context.previewData === 'object' && 'currentPreviewApiKey' in context.previewData
-    ? context.previewData.currentPreviewApiKey as string
-    : undefined;
+  const previewApiKey = getPreviewApiKeyFromPreviewData(context.previewData);
 
   const solution = await getSolutionDetail({ envId, previewApiKey }, slug, !!context.preview);
   const siteMenu = await getSiteMenu({ envId, previewApiKey }, !!context.preview);
